requirements.txt
torch
pandas
transformers
pinecone-client
python-dotenv
pydantic
groq
PyQt5
timescale-vector